Reorganize Buttons' folder hierarchy
