+++ src/components/PageSwitch.tsx
@@ -3,7 +3,7 @@ import { ChevronRight, ChevronLeft } from 'react-feather'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 import useQueryParams from '../hooks/useQueryParams'
-import { TextButton } from './Buttons'
+import { TextButton } from './Buttons/TextButton'
 
 const PageSwitch = ({
   elementsPerPage = 20,
